\name{DemoModel}
\alias{DemoModel}
\title{DemoModel}
\description{
  Simple example of an implementatiopn of a predictive
  model used in the package demo example and used as a
  baseline to evaluate performance of other models. This
  model builds a random survival forest using clinical
  covariates known to relate to breast cancer survival,
  gene expression and copy number data. It also imputes
  missing covariates. NAs should be expected in all
  clinical covariates, in both training and test data sets.
}
\author{
  Erhan Bilal
}

